requirements.txt
feedparser
requests
cryptography
streamlit
PyYAML
hydra-core 
omegaconf
python-dotenv
pytest
